move cpu/cuda related files to coresponding cpu/cuda folder (#3668)
