+++ orttraining/orttraining/training_ops/cpu/cpu_training_kernels.cc
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-#include "orttraining/training_ops/cpu_training_kernels.h"
+#include "orttraining/training_ops/cpu/cpu_training_kernels.h"
 #include "core/graph/constants.h"
 
 namespace onnxruntime {
